// api/generated/model/reportSubmissionDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { ReportRequestDto } from './reportRequestDto';
import type { ReportSubmissionDtoAnswers } from './reportSubmissionDtoAnswers';
import type { ReportSubmissionDtoTrainerFieldComments } from './reportSubmissionDtoTrainerFieldComments';

export interface ReportSubmissionDto {
  /** @nullable */
  _id?: string | null;
  /** @nullable */
  reportRequestId?: string | null;
  /** @nullable */
  traineeId?: string | null;
  submittedAt?: string;
  /** @nullable */
  answers?: ReportSubmissionDtoAnswers;
  /** @nullable */
  trainerOverallComment?: string | null;
  /** @nullable */
  trainerFieldComments?: ReportSubmissionDtoTrainerFieldComments;
  request?: ReportRequestDto;
}